+++ R/as_measure.R
@@ -10,13 +10,13 @@
 #'
 #' @return [Measure].
 #' @export
-as_measure = function(x, task_type = NULL, ...) { # nolint
+as_measure = function(x, task_type = NULL, clone = FALSE, ...) { # nolint
   UseMethod("as_measure")
 }
 
 #' @export
 #' @rdname as_measure
-as_measure.NULL = function(x, task_type = NULL, ...) { # nolint
+as_measure.NULL = function(x, task_type = NULL, clone = FALSE, ...) { # nolint
   assert_empty_ellipsis(...)
   default_measures(task_type)[[1L]]
 }
@@ -30,24 +30,24 @@ as_measure.Measure = function(x, task_type = NULL, clone = FALSE, ...) { # nolin
 
 #' @export
 #' @rdname as_measure
-as_measures = function(x, task_type = NULL, ...) { # nolint
+as_measures = function(x, task_type = NULL, clone = FALSE, ...) { # nolint
   UseMethod("as_measures")
 }
 
 #' @export
 #' @rdname as_measure
-as_measures.default = function(x, task_type = NULL, ...) { # nolint
-  list(as_measure(x, task_type = task_type, ...))
+as_measures.default = function(x, task_type = NULL, clone = FALSE, ...) { # nolint
+  list(as_measure(x, task_type = task_type, clone = clone, ...))
 }
 
 #' @export
 #' @rdname as_measure
-as_measures.NULL = function(x, task_type = NULL, ...) { # nolint
+as_measures.NULL = function(x, task_type = NULL, clone = FALSE, ...) { # nolint
   default_measures(task_type)
 }
 
 #' @export
 #' @rdname as_measure
-as_measures.list = function(x, task_type = NULL, ...) { # nolint
-  lapply(x, as_measure, task_type = NULL, ...)
+as_measures.list = function(x, task_type = NULL, clone = FALSE, ...) { # nolint
+  lapply(x, as_measure, task_type = NULL, clone = clone, ...)
 }