+++ man-roxygen/param_callbacks.R
@@ -1,0 +1,3 @@
+#' @param callbacks (List of [mlr3misc::Callback])\cr
+#' Callbacks to be executed during the resampling process.
+#' See [CallbackResample] and [ContextResample] for details.